Fix a bug in displaying :help text at top of web REPL history
The previous implemenation was adding a fake input and output
to the history and then removing just the input from the DOM.
We also removed the entry from the `inputHistory` array, to
prevent it appearing in the panel when you hit the UP arrow
enough times.
However this introduces a mismatch between the number of items
in `inputHistory` array and the number of `.history-item` divs
in the DOM.
This caused each REPL output to appear *above* the user input
that caused it! :upside_down_face:

The fix is to just make all of this separate from history!
We just make a separate call into Rust to get the initial :help.

The original idea was to share code between the initial :help
and the normal REPL operation, but they turn out to be only
similar, not the same.